requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
razorpay>=1.4.1

cloudinary>=1.36.0
